requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
structlog==24.1.0
python-json-logger==2.0.7

# Async and Networking
aiohttp==3.9.1
aiofiles==23.2.1
asyncpg==0.29.0
websockets==12.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b

# Database
sqlalchemy==2.0.41
alembic==1.16.4
redis==6.2.0
asyncpg==0.29.0

# API and Authentication
httpx==0.27.2
pyjwt==2.8.0
cryptography==41.0.7

# UI Development
PyQt5==5.15.10
pyqtgraph==0.13.3
matplotlib==3.8.2
plotly==5.18.0

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
mypy==1.7.1
black==23.12.0
isort==5.13.2
flake8==6.1.0
pylint==3.0.3

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Performance and Profiling
memory-profiler==0.61.0
line-profiler==4.1.2
py-spy==0.3.14

# Utilities
click==8.1.7
tqdm==4.66.1
tenacity==8.2.3
cachetools==5.3.2